requirements.txt
django
django-import-export
requests
notify-py
